pyproject.toml
[project]
name = "tinymolty"
version = "0.1.0"
description = "A cross-platform asyncio bot for moltbook.com with configurable AI personality."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "openai>=1.0",
    "google-genai>=1.0",
    "textual>=1.0",
    "keyring>=25.0",
    "tomli-w>=1.0",
]

[project.scripts]
tinymolty = "__main__:main"

[project.urls]
Homepage = "https://github.com/herrkaefer/tinymolty"
Repository = "https://github.com/herrkaefer/tinymolty"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
